format *all* the files

+++ test/mocks/graphics_mock.h
@@ -1,14 +1,16 @@
 #ifndef __GRAPHICS_MOCK_H__
 #define __GRAPHICS_MOCK_H__
 
-#include <gmock/gmock.h>
 #include "../../src/core/graphics.h"
+#include <gmock/gmock.h>
 
 class GraphicsMock : public Graphics {
 public:
-    MOCK_METHOD5 (drawTile, void (int y, int x, unsigned int tile, Color fg, Color bg));
-    MOCK_METHOD5 (drawString, unsigned int (int y, int x, const char* s, Color fg, Color bg));
-    MOCK_METHOD4 (drawBorder, void(int y, int x, int height, int width));
+    MOCK_METHOD5(drawTile,
+                 void(int y, int x, unsigned int tile, Color fg, Color bg));
+    MOCK_METHOD5(drawString,
+                 unsigned int(int y, int x, const char* s, Color fg, Color bg));
+    MOCK_METHOD4(drawBorder, void(int y, int x, int height, int width));
 };
 
 #endif
